feat: 가게 등록, 삭제 동시성 문제 해결

+++ shop-service/src/main/java/com/node5/shopservice/shop/infrastructure/kafka/consumer/MemberRoleChangeFailedConsumer.java
@@ -1,6 +1,7 @@
 package com.node5.shopservice.shop.infrastructure.kafka.consumer;
 
-import com.node5.common.event.ShopDeletionFailedEvent;
+import com.node5.common.event.MemberRoleChangeFailedEvent;
+import com.node5.common.event.MemberRoleChangeSagaType;
 import com.node5.shopservice.shop.application.ShopService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -11,16 +12,20 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ShopDeletionFailedConsumer {
+public class MemberRoleChangeFailedConsumer {
 
     private final ShopService shopService;
 
     @KafkaListener(
-            topics = "${kafka.topics.shop-deletion-failed}",
+            topics = "${kafka.topics.member-role-change-failed}",
             containerFactory = "retryKafkaListenerContainerFactory"
     )
-    public void consume(ShopDeletionFailedEvent event, Acknowledgment ack) {
-        shopService.deleteShopFailed(event);
+    public void consume(MemberRoleChangeFailedEvent event, Acknowledgment ack) {
+        if (event.sagaType() == MemberRoleChangeSagaType.SHOP_REGISTRATION) {
+            shopService.registerShopFailed(event);
+        } else if (event.sagaType() == MemberRoleChangeSagaType.SHOP_DELETION) {
+            shopService.deleteShopFailed(event);
+        }
         ack.acknowledge();
     }
 }
